refactor: 막차 조회 로직 분리
