+++ src/main/kotlin/com/deepromeet/atcha/transit/domain/TransitService.kt
@@ -1,29 +1,18 @@
 package com.deepromeet.atcha.transit.domain
 
 import com.deepromeet.atcha.location.domain.Coordinate
-import com.deepromeet.atcha.transit.api.response.LastRouteLeg
 import com.deepromeet.atcha.transit.api.response.LastRoutesResponse
 import com.deepromeet.atcha.transit.exception.TransitException
-import com.deepromeet.atcha.transit.infrastructure.client.tmap.TMapTransitClient
-import com.deepromeet.atcha.transit.infrastructure.client.tmap.request.TMapRouteRequest
-import com.deepromeet.atcha.transit.infrastructure.client.tmap.response.Itinerary
-import com.deepromeet.atcha.transit.infrastructure.client.tmap.response.Leg
 import com.deepromeet.atcha.user.domain.UserReader
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.withContext
 import org.springframework.stereotype.Service
-import java.time.Duration
-import java.time.LocalDate
 import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
-import java.util.UUID
 
 @Service
 class TransitService(
-    private val tMapTransitClient: TMapTransitClient,
     private val taxiFareFetcher: TaxiFareFetcher,
     private val busManager: BusManager,
     private val subwayManager: SubwayManager,
@@ -33,7 +22,8 @@ class TransitService(
     private val lastRouteReader: LastRouteReader,
     private val lastRouteAppender: LastRouteAppender,
     private val lastRouteIndexReader: LastRouteIndexReader,
-    private val lastRouteIndexAppender: LastRouteIndexAppender
+    private val lastRouteIndexAppender: LastRouteIndexAppender,
+    private val lastRouteOperations: LastRouteOperations
 ) {
     fun init() {
         subwayStationBatchAppender.appendAll()
@@ -101,20 +91,16 @@ class TransitService(
         regionIdentifier.identify(start)
         regionIdentifier.identify(end)
 
-        val today = LocalDate.now()
-        val dateFormatter = DateTimeFormatter.ofPattern("yyyyMMdd")
-        val baseDate = today.format(dateFormatter) + "2300"
+        // 막차 조회
+        val allRoutes = lastRouteOperations.getItineraries(start, end)
+        val deduplicatedRoutes = lastRouteOperations.filterAndDeduplicateItineraries(allRoutes)
 
-        val allRoutes = getItineraries(baseDate, start, end)
-        val deduplicatedRoutes = filterAndDeduplicateItineraries(allRoutes)
-
-        // 경로 막차 계산
         val lastRoutesResponses =
             coroutineScope {
                 deduplicatedRoutes
                     .map { route ->
                         async(Dispatchers.Default) {
-                            calculateRoute(route)
+                            lastRouteOperations.calculateRoute(route)
                         }
                     }
                     .awaitAll()
@@ -122,321 +108,13 @@ class TransitService(
             }
 
         val now = LocalDateTime.now()
-        val filteredRoutes =
-            lastRoutesResponses.filter { response ->
-                val departureDateTime = LocalDateTime.parse(response.departureDateTime)
-                departureDateTime.isAfter(now)
-            }
+        val filteredRoutes = lastRouteOperations.getFilteredRoutes(lastRoutesResponses, now)
 
         saveRouteIdsByStartEnd(start, end, filteredRoutes.map { it.routeId })
         saveRoutesToCache(filteredRoutes)
 
-        return sortedByMinTransfer(filteredRoutes)
+        return lastRouteOperations.sortedByMinTransfer(filteredRoutes)
     }
-
-    private suspend fun getItineraries(
-        baseDate: String,
-        start: Coordinate,
-        end: Coordinate
-    ): List<Itinerary> {
-        val response =
-            withContext(Dispatchers.IO) {
-                tMapTransitClient.getRoutes(
-                    TMapRouteRequest(
-                        startX = start.lon.toString(),
-                        startY = start.lat.toString(),
-                        endX = end.lon.toString(),
-                        endY = end.lat.toString(),
-                        count = 10,
-                        searchDttm = baseDate
-                    )
-                )
-            }
-
-        response.result?.let { result ->
-            when (result.status) {
-                11 -> throw TransitException.DistanceTooShort
-                else -> throw TransitException.ServiceAreaNotSupported
-            }
-        }
-
-        return response.metaData?.plan?.itineraries ?: throw TransitException.TransitApiError
-    }
-
-    private fun filterAndDeduplicateItineraries(itineraries: List<Itinerary>): List<Itinerary> {
-        return itineraries.filterNot { itinerary ->
-            val transitModes = itinerary.legs.filter { it.mode == "SUBWAY" || it.mode == "BUS" }
-            val busCountExcludingFirst = transitModes.drop(1).count { it.mode == "BUS" }
-            val hasValidModes =
-                itinerary.legs.any {
-                    it.mode == "WALK" ||
-                        (it.mode == "SUBWAY" && it.route != null && !it.route.contains("(급행)")) ||
-                        it.mode == "BUS"
-                }
-            !hasValidModes || (transitModes.size >= 3 && busCountExcludingFirst >= 2) || transitModes.size >= 5
-        }.associateBy { itinerary ->
-            itinerary.legs.joinToString("|") { leg ->
-                "${leg.start.name}-${leg.end.name}-${leg.route ?: ""}"
-            }
-        }.values.toList()
-    }
-
-    private suspend fun calculateRoute(route: Itinerary): LastRoutesResponse? {
-        // 1. 경로 내 대중교통 별 막차 시간 조회
-        val calculatedLegs = calculateLegLastArriveDateTimes(route.legs) ?: return null
-        // 2. 도보 시간 조정  - 모든 도보는 2분씩 더해준다.
-        val adjustedWalkLegs = increaseWalkTime(calculatedLegs)
-        // 3. 막차 시간 기준, 경로 내 대중교통 탑승 가능 여부 확인
-        val adjustedLegs = adjustTransitDepartureTimes(adjustedWalkLegs)
-        // 4. 유효하지 않는 경로 제거 (leg.departureTime=null 인 경우)
-        if (adjustedLegs.any { leg ->
-                (leg.mode == "SUBWAY" || leg.mode == "BUS") &&
-                    (leg.departureDateTime == null || leg.departureDateTime == "null")
-            }
-        ) {
-            return null
-        }
-        // 5. 출발 시간 계산
-        val departureDateTime = calculateDepartureDateTime(adjustedLegs)
-        // 6. 총 소요 시간 계산
-        val totalTime = calculateTotalTime(adjustedLegs, departureDateTime)
-
-        return LastRoutesResponse(
-            routeId = UUID.randomUUID().toString(),
-            departureDateTime = departureDateTime.toString(),
-            totalTime = totalTime.toInt(),
-            totalWalkTime = route.totalWalkTime,
-            transferCount = route.transferCount,
-            totalDistance = route.totalDistance,
-            pathType = 0,
-            legs = adjustedLegs
-        )
-    }
-
-    private suspend fun calculateLegLastArriveDateTimes(legs: List<Leg>): List<LastRouteLeg>? {
-        val calculatedLegs =
-            coroutineScope {
-                legs.map { leg ->
-                    async(Dispatchers.IO) {
-                        val departureDateTime =
-                            when (leg.mode) {
-                                "SUBWAY" ->
-                                    getLastTime(
-                                        SubwayLine.fromRouteName(leg.route!!),
-                                        leg.start.name,
-                                        leg.end.name
-                                    )?.departureTime?.toString()
-
-                                "BUS" ->
-                                    getBusArrivalInfo(
-                                        leg.route!!.split(":")[1],
-                                        leg.start.name.removeSuffix(),
-                                        Coordinate(leg.start.lat, leg.start.lon)
-                                    )?.lastTime?.toString()
-
-                                else -> null
-                            }
-                        leg.toLeg(departureDateTime)
-                    }
-                }.awaitAll().toList()
-            }
-
-        // null이 섞여있는 leg가 하나라도 있다면, 전체 경로 자체를 제거
-        return if (calculatedLegs.any {
-                (it.mode == "BUS" || it.mode == "SUBWAY") &&
-                    (it.departureDateTime == null || it.departureDateTime == "null")
-            }
-        ) {
-            null
-        } else {
-            calculatedLegs
-        }
-    }
-
-    private fun increaseWalkTime(legs: List<LastRouteLeg>): List<LastRouteLeg> {
-        return legs.mapIndexed { index, currentLeg ->
-            val nextLeg = legs.getOrNull(index + 1)
-            if (currentLeg.mode == "WALK" && (nextLeg?.mode == "SUBWAY" || nextLeg?.mode == "BUS")) {
-                currentLeg.copy(sectionTime = currentLeg.sectionTime + 120)
-            } else {
-                currentLeg
-            }
-        }
-    }
-
-    private fun adjustTransitDepartureTimes(legs: List<LastRouteLeg>): List<LastRouteLeg> {
-        val adjustedLegs = legs.toMutableList()
-        if (adjustedLegs.any { it.mode != "WALK" && it.departureDateTime == null }) return adjustedLegs
-
-        val transitLegs = adjustedLegs.withIndex().filter { it.value.mode != "WALK" }
-
-        if (transitLegs.isEmpty()) return adjustedLegs
-
-        // 1. 대중교통 기준 가장 빠른 막차 시간 찾기
-        val earliestTransitLeg = transitLegs.minBy { LocalDateTime.parse(it.value.departureDateTime!!) }
-
-        var isAllRideable = true
-        var lastUnrideableIndex: Int? = null
-
-        // 2. 가장 빠른 출발 시간을 기준으로 뒤에 있는 대중교통 탑승 가능 여부 확인
-        for (i in earliestTransitLeg.index until adjustedLegs.lastIndex) {
-            val currentLeg = adjustedLegs[i]
-            if (currentLeg.mode == "WALK") continue
-
-            // 2-1. 출발 시간 + 소요 시간
-            var currentLegAvailableTime =
-                LocalDateTime.parse(currentLeg.departureDateTime!!).plusSeconds(currentLeg.sectionTime.toLong())
-
-            // 2-2. 2-1 결과 시간과 다음 대중교통 출발 시간 비교 -> 탑승 가능 여부 확인
-            var nextIndex = i + 1
-            while (nextIndex <= adjustedLegs.lastIndex && adjustedLegs[nextIndex].mode == "WALK") {
-                currentLegAvailableTime =
-                    currentLegAvailableTime.plusSeconds(adjustedLegs[nextIndex].sectionTime.toLong())
-                nextIndex++
-            }
-
-            if (nextIndex > adjustedLegs.lastIndex) break
-
-            val nextLeg = adjustedLegs[nextIndex]
-            val nextLegArriveTime = LocalDateTime.parse(nextLeg.departureDateTime!!)
-
-            if (currentLegAvailableTime.isAfter(nextLegArriveTime)) {
-                isAllRideable = false
-                lastUnrideableIndex = nextIndex
-            }
-        }
-
-        // 3. 기준점 설정 : 가장 빠른 출발 시간 기준 or 탑승 불가한 마지막 대중교통
-        val adjustBaseIndex = if (isAllRideable) earliestTransitLeg.index else lastUnrideableIndex!!
-
-        // 4. 기준점 앞쪽 시간 재조정
-        var adjustBaseTime = LocalDateTime.parse(adjustedLegs[adjustBaseIndex].departureDateTime!!) ?: null
-        for (i in adjustBaseIndex - 1 downTo 0) {
-            val leg = adjustedLegs[i]
-
-            if (adjustBaseTime == null) {
-                adjustedLegs[i] = leg.copy(departureDateTime = null)
-                continue
-            }
-
-            if (leg.mode == "WALK") {
-                adjustBaseTime = adjustBaseTime.minusSeconds(leg.sectionTime.toLong())
-                continue
-            }
-
-            val adjustedDepartureTime = adjustBaseTime.minusSeconds(leg.sectionTime.toLong())
-            val calculateBoardingTime = calculateBoardingTime(leg, adjustedDepartureTime, TimeDirection.BEFORE)
-            adjustedLegs[i] = leg.copy(departureDateTime = calculateBoardingTime?.toString())
-            adjustBaseTime = calculateBoardingTime?.let { LocalDateTime.parse(it.toString()) }
-        }
-
-        // 5. 기준점 뒤쪽 시간 재조정
-        adjustBaseTime =
-            LocalDateTime.parse(adjustedLegs[adjustBaseIndex].departureDateTime!!)
-                .plusSeconds(adjustedLegs[adjustBaseIndex].sectionTime.toLong())
-        for (i in adjustBaseIndex + 1 until adjustedLegs.size) {
-            val leg = adjustedLegs[i]
-
-            if (adjustBaseTime == null) {
-                adjustedLegs[i] = leg.copy(departureDateTime = null)
-                continue
-            }
-
-            if (leg.mode == "WALK") {
-                adjustBaseTime = adjustBaseTime.plusSeconds(leg.sectionTime.toLong())
-                continue
-            }
-
-            val adjustedDepartureTime = adjustBaseTime
-            val calculateBoardingTime = calculateBoardingTime(leg, adjustedDepartureTime, TimeDirection.AFTER)
-            adjustedLegs[i] = leg.copy(departureDateTime = calculateBoardingTime.toString())
-            adjustBaseTime = calculateBoardingTime?.plusSeconds(leg.sectionTime.toLong())
-        }
-        return adjustedLegs
-    }
-
-    fun calculateBoardingTime(
-        leg: LastRouteLeg,
-        adjustedDepartureTime: LocalDateTime,
-        direction: TimeDirection
-    ): LocalDateTime? =
-        when (leg.mode) {
-            "SUBWAY" -> {
-                val subwayLine = SubwayLine.fromRouteName(leg.route!!)
-                val routes = subwayManager.getRoutes(subwayLine)
-                val startStation = subwayManager.getStation(subwayLine, leg.start.name)
-                val endStation = subwayManager.getStation(subwayLine, leg.end.name)
-
-                subwayManager.getTimeTable(startStation, endStation, routes)
-                    ?.findNearestTime(adjustedDepartureTime, direction)?.departureTime
-            }
-
-            "BUS" ->
-                getBusArrivalInfo(
-                    leg.route!!.split(":")[1],
-                    leg.start.name.removeSuffix(),
-                    Coordinate(leg.start.lat, leg.start.lon)
-                )?.getNearestTime(adjustedDepartureTime, direction)
-
-            else -> null
-        }
-
-    fun calculateDepartureDateTime(legs: List<LastRouteLeg>): LocalDateTime {
-        val firstTransitIndex = legs.indexOfFirst { it.mode != "WALK" }
-        val firstTransit = legs[firstTransitIndex]
-        val departureDateTime = LocalDateTime.parse(firstTransit.departureDateTime!!)
-        val totalWalkTime =
-            if (firstTransitIndex > 0) {
-                legs.subList(0, firstTransitIndex).filter { it.mode == "WALK" }.sumOf { it.sectionTime.toLong() }
-            } else {
-                0
-            }
-
-        return departureDateTime.minusSeconds(totalWalkTime)
-    }
-
-    private fun calculateTotalTime(
-        adjustedLegs: List<LastRouteLeg>,
-        departureDateTime: LocalDateTime
-    ): Long {
-        val lastTransitIndex = adjustedLegs.indexOfLast { it.mode != "WALK" }
-        val lastTransit = adjustedLegs[lastTransitIndex]
-
-        val lastTransitDepartureTime = LocalDateTime.parse(lastTransit.departureDateTime!!)
-        var arrivalTime = lastTransitDepartureTime.plusSeconds(lastTransit.sectionTime.toLong())
-
-        val totalWalkTime =
-            adjustedLegs.drop(lastTransitIndex + 1) // 마지막 대중교통 이후 구간
-                .filter { it.mode == "WALK" }.sumOf { it.sectionTime.toLong() }
-
-        arrivalTime = arrivalTime.plusSeconds(totalWalkTime)
-
-        // 4. 총 소요 시간 계산 (초 단위)
-        return Duration.between(departureDateTime, arrivalTime).seconds
-    }
-
-    private fun Leg.toLeg(departureDateTime: String?): LastRouteLeg {
-        return LastRouteLeg(
-            distance = this.distance,
-            sectionTime = this.sectionTime,
-            mode = this.mode,
-            departureDateTime = departureDateTime,
-            route = this.route,
-            type = this.type.toString(),
-            service = this.service.toString(),
-            start = this.start,
-            end = this.end,
-            passStopList = this.passStopList?.stationList,
-            step = this.steps,
-            passShape = this.passShape?.linestring
-        )
-    }
-
-    // 정렬 : 환승-시간 순으로 정렬로 고정
-    private fun sortedByMinTransfer(routes: List<LastRoutesResponse>) =
-        routes.sortedWith(
-            compareBy({ it.transferCount }, { it.totalTime })
-        )
 
     private fun saveRouteIdsByStartEnd(
         start: Coordinate,
@@ -449,6 +127,4 @@ class TransitService(
     private fun saveRoutesToCache(routes: List<LastRoutesResponse>) {
         routes.forEach { route -> lastRouteAppender.append(route) }
     }
-
-    private fun String.removeSuffix(): String = this.replace("(중)", "").trim()
 }
